Support for projection functors in the new core (#233)
* Take optional projection functors in ManualTask

* Replace SymbolicFunctor with SymbolicPoint

* API to retrieve child stores

* Missing files

* Binding for SymbolicExpr

* Add a projection parameter to ManualTask.add_{input,output,reduction}

* Missing nodiscard on a to_string method

* Support launch domains with non-zero lower bounds for manual tasks

* Turn off interference checks for ManualTasks using custom projections

* Address review comments

* Minor fix to make clang-tidy happy

* More clang-tidy fixes

* Fix compile error

+++ src/core/operation/detail/reduce.h
@@ -33,7 +33,7 @@ class Reduce final : public Operation {
          std::shared_ptr<LogicalStore> out_store,
          int64_t task_id,
          uint64_t unique_id,
-         int64_t radix,
+         int32_t radix,
          mapping::detail::Machine&& machine);
 
  public:
@@ -45,7 +45,7 @@ class Reduce final : public Operation {
   [[nodiscard]] std::string to_string() const override;
 
  private:
-  int64_t radix_{};
+  int32_t radix_{};
   const Library* library_{};
   int64_t task_id_{};
   std::shared_ptr<LogicalStore> input_{};
